Introducing MobArena Labs.
This commit introduces the concept of MobArena Labs as well as the first
explicitly experimental feature in the form of custom entity type lists
for the cleanup procedure.

MobArena Labs is an attempt at a compromise between the otherwise rigid
approach to functionality due to maintainability concerns and some of
the esoteric feature requests we see on Github. The compromise manifests
itself as follows:

- All experimental functionality is opt-in and set up primarily in a new
  file, `labs.yml`, which must be created manually. Nothing is provided
  for users by default, so the "I know what I'm doing"-attitude required
  for this functionality to work hopefully helps convey the user-facing
  downside; that experimental functionality is much more likely to break
  between releases, and that compatibility issues are ignored.

- The other side of the coin is that we can try new things as long as
  they are _local to MobArena_, i.e. no external dependencies. Some of
  the esoteric and niche feature requests we see on Github could make it
  into the plugin as experimental features. If enough servers make use
  of them and provide feedback, they could end up as regular features
  with stability and robustness concerns on par with core parts of the
  plugin. We use bStats to help visualize feature usage (or lack thereof
  so we can remove unused functionality if it is problematic).

The first experimental feature is the ability to customize the list of
entity types that MobArena looks for when clearing the arena region of
residual entities such as arrows and experience orbs. This is called the
Housekeeper, and it has its own section in `labs.yml`. It can be toggled
on or off with the `enabled` flag, and the list of entities is specified
with the `entities` list. Because the original code was hardcoded in the
ArenaImpl class, no effort has been made to make the Housekeeper and its
settings per-arena configurable.

Closes #667